# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Canela Ceylon SCM API
  version: 1.0.0
  description: >-
    Contract-first API for AI-integrated supply chain and blockchain traceability
    operations.
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /api/v1/blockchain-registration/{batchId}:
    post:
      tags: [blockchain-registration]
      summary: Register batch metadata on Sepolia
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Registration submitted
  /api/v1/verify/{batchId}:
    get:
      tags: [public-traceability-verification]
      summary: Public traceability lookup
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification data
...